% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{htmlForCalibrationHistory}
\alias{htmlForCalibrationHistory}
\title{History HTML for calibration progress visualisation}
\usage{
htmlForCalibrationHistory(n, current = "ucCurrent.html")
}
\arguments{
\item{n}{number of plots}

\item{current}{default: "ucCurrent.html"}
}
\description{
History HTML for calibration progress visualisation
}
