% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{htmlForCalibrationPage}
\alias{htmlForCalibrationPage}
\title{Main HTML for calibration progress visualisation}
\usage{
htmlForCalibrationPage(
  htmlHist = "ucHistory.html",
  htmlCurr = "ucCurrent.html"
)
}
\arguments{
\item{htmlHist}{default: "ucHistory.html"}

\item{htmlCurr}{default: "ucCurrent.html"}
}
\description{
Main HTML for calibration progress visualisation
}
