% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{htmlForCalibrationState}
\alias{htmlForCalibrationState}
\title{Content HTML for calibration progress visualisation}
\usage{
htmlForCalibrationState(refresh = 1)
}
\arguments{
\item{refresh}{\code{refresh} time in seconds}
}
\description{
Content HTML for calibration progress visualisation
}
