% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucConf}
\alias{ucConf}
\title{Configure UCODE run}
\usage{
ucConf(
  general = ucConfGeneral(),
  parameters = ucDefaultConfParameters(),
  weights = ucDefaultConfWeights()
)
}
\arguments{
\item{general}{configuration of control \code{parameters} as retrieved by
\code{\link{ucConfGeneral}}.}

\item{parameters}{data.frame with each row representing a parameter
configuration as e.g. retrieved by \code{\link{ucConfParameter}}. For
possible parameter names see \code{\link{ucSetPlaceholder}}}

\item{weights}{Matrix of \code{weights} as e.g. retrieved by
\code{\link{ucDefaultConfWeights}}. Matix of \code{weights} is used when
calculating overall fitness. Row names = names of fitness indicators out of
the set of c("ME", "MAE", "MSE", "RMSE", "NRMSE", "PBIAS", "RSR", "rSD",
"NSE", "mNSE", "rNSE", "d", "md", "rd", "cp", "r", "R2", "bR2", "KGE,
"VE"). Column names = well names ("PW" = pumping well, observation wells
named as in parameter setting). See example.}
}
\description{
Configure UCODE run
}
\seealso{
\code{\link{ucPrepCalib}}
}
