% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucConf_Model_Command_Lines}
\alias{ucConf_Model_Command_Lines}
\title{ucConf Model Command Lines}
\usage{
ucConf_Model_Command_Lines(bfile)
}
\arguments{
\item{bfile}{name of batch file to be invoked by UCODE}
}
\description{
lines for input block \emph{Model_Command_Lines}
}
