% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucConf_Observation_Data}
\alias{ucConf_Observation_Data}
\title{Configuration of Observation Data}
\usage{
ucConf_Observation_Data(weights)
}
\arguments{
\item{weights}{Matix of \code{weights} used when calculating overall fitness.
Row names = names of fitness indicators out of the set of c("ME", "MAE",
"MSE", "RMSE", "NRMSE", "PBIAS", "RSR", "rSD", "NSE", "mNSE", "rNSE",
"d", "md", "rd", "cp", "r", "R2", "bR2", "KGE, "VE"). Column names = well
names ("PW" = pumping well, observation wells named as in parameter
setting). For an example see \code{\link{ucConf}}}
}
\description{
Lines for input block \emph{Observation_Data}
}
