% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucConf_Parameter_Data}
\alias{ucConf_Parameter_Data}
\title{Configuration of Parameter Data}
\usage{
ucConf_Parameter_Data(confParameters)
}
\arguments{
\item{confParameters}{data frame representing table as required in Parameter_Data block of
UCODE input file. , for example see \code{\link{ucDefaultConfParameters}}}
}
\description{
Lines for input block \emph{Parameter_Data}
}
