% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucConf_Reg_GN_Controls}
\alias{ucConf_Reg_GN_Controls}
\title{ucConf Reg GN Controls}
\usage{
ucConf_Reg_GN_Controls(maxIter)
}
\arguments{
\item{maxIter}{Maximum number of parameter-estimation iterations allowed before stopping}
}
\description{
lines for input block \emph{UCODE_Control_Data}
}
