% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucConfGeneral}
\alias{ucConfGeneral}
\title{General settings in UCODE configuration}
\usage{
ucConfGeneral(gof.digits = 5, maxIter = 100)
}
\arguments{
\item{gof.digits}{Number of digits to be used in goodness of fit values.
Default: 5}

\item{maxIter}{Maximum number of parameter-estimation iterations allowed
before stopping}
}
\description{
General settings in UCODE configuration
}
\seealso{
\code{\link{ucConf}}
}
