% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucConfParameter}
\alias{ucConfParameter}
\title{Create parameter for UCODE configuration}
\usage{
ucConfParameter(
  paramname,
  startvalue = 1e+38,
  lowerValue = -1e+38,
  upperValue = 1e+38,
  lowerConstraint = 0,
  upperConstraint = 1,
  constrain = "yes",
  adjustable = "yes",
  maxChange = 2,
  perturbAmt = 0.5
)
}
\arguments{
\item{paramname}{Parameter name (up to 12 characters; not case sensitive) - a
character string that is used in a template file or in an equation of a
derived parameter in the Derived_Parameters input block. Each parameter
name needs to be unique and can not be the same as any parameter name
defined in the Derived_Parameter input block. For possible parameter names
see \code{\link{ucSetPlaceholder}}}

\item{startvalue}{Starting parameter value. Default=A huge real number. The
huge real number is obtained for the computer being used and commonly is
about 1e38.}

\item{lowerValue}{Smallest reasonable value for this parameter. Default=
-(Huge real number). In absolute value, commonly about -1e38.}

\item{upperValue}{Largest reasonable value for this parameter. Default=
+(Huge real number). Commonly about +1e38.}

\item{lowerConstraint}{Lower limit of considered parameter values. Default
(HS): 0}

\item{upperConstraint}{Upper limit of considered parameter values. Default
(HS): 1}

\item{constrain}{yes: \code{constrain} parameter values using LowerConstraint
and UpperConstraint. Default (HS): yes.}

\item{adjustable}{yes: change this value as needed depending on the purpose
of the UCODE_2005 run defined in the UCODE_Control_Data input file. no:
leave the value of this parameter unchanged. Default (HS): "yes".}

\item{maxChange}{Maximum fractional parameter change allowed between
parameter iterations. Default=2.0.}

\item{perturbAmt}{Fractional amount of parameter value to perturb to
calculate sensitivity. Commonly 0.01 to 0.10. Default (HS): 0.5. See
discussion in Chapter 3.}
}
\value{
data.frame containing all argument values with column names equalling
function argument names
}
\description{
Create parameter for UCODE configuration
}
\examples{
  # Use rbind to create a data.frame containing parameter information as
  # required by ucConf:
  cpara <- rbind(
    ucConfParameter(paramname = "HKR", startvalue = 2e-5, 
                    lowerConstraint = 1e-5, upperConstraint = 1e-1,
                    maxChange = 1, perturbAmt = 0.5, 
                    constrain = "yes", adjustable = "yes"),
    ucConfParameter(paramname = "HKZ", startvalue = 2e-7, 
                    lowerConstraint = 1e-7, upperConstraint = 1e-3,
                    maxChange = 1, perturbAmt = 0.5,
                    constrain = "yes", adjustable = "yes"))  
  
  # set this parameter setting in default UCODE configuration
  ucConf(parameters = cpara)

}
\seealso{
\code{\link{ucConf}, \link{ucDefaultConfParameters}}
}
