% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucDefaultConfParameters}
\alias{ucDefaultConfParameters}
\title{Default parameter definition for UCODE configuration}
\usage{
ucDefaultConfParameters()
}
\value{
data.frame with each row representing a model parameter
}
\description{
Default parameter definition for UCODE configuration
}
\seealso{
\code{\link{ucConfParameter}, \link{ucConf}}
}
