% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucDefaultConfWeights}
\alias{ucDefaultConfWeights}
\title{Default weight matrix}
\usage{
ucDefaultConfWeights(
  wells = wtConfiguredWellnames(wtDefaultConfiguration()),
  gofs = c("RMSE", "NSE", "mNSE")
)
}
\arguments{
\item{wells}{well names, default:
\code{wtConfiguredWellnames(wtDefaultConfiguration())}}

\item{gofs}{default: \code{c("RMSE", "NSE", "mNSE")}}
}
\value{
Matrix of weights used when calculating overall fitness.
Row names = names of fitness indicators out of the set of c("ME", "MAE",
"MSE", "RMSE", "NRMSE", "PBIAS", "RSR", "rSD", "NSE", "mNSE", "rNSE",
"d", "md", "rd", "cp", "r", "R2", "bR2", "KGE, "VE"). Column names = well
names ("PW" = pumping well, observation \code{wells} named as in parameter
setting).
}
\description{
Weights are given for each pair of well name (pumping well \emph{PW} or
observation \code{wells}) and name of goodness of fit (gof)-function.
}
\seealso{
\code{\link{ucConf}}
}
