% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucInpFileLines}
\alias{ucInpFileLines}
\title{Input File Lines}
\usage{
ucInpFileLines(uconf, command)
}
\arguments{
\item{uconf}{UCODE configuration as retrieved by \code{\link{ucConf}}}

\item{command}{name of batch file that is invoked by UCODE with each iteration}
}
\description{
Lines for UCODE main input file
}
