% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucPlotPng}
\alias{ucPlotPng}
\title{Plot WTAQ results to png file}
\usage{
ucPlotPng(fname, wtaqResult, i)
}
\arguments{
\item{fname}{fname}

\item{wtaqResult}{wtaqResult}

\item{i}{i}
}
\value{
????
}
\description{
Plot WTAQ results to png file
}
