% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucReadSummary}
\alias{ucReadSummary}
\title{Read UCODE summary file}
\usage{
ucReadSummary(sfile)
}
\arguments{
\item{sfile}{path to summary file}
}
\value{
data frame, ordered by ITER
}
\description{
Read UCODE summary file
}
