% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucRunParallel}
\alias{ucRunParallel}
\title{Run main calibration loop}
\usage{
ucRunParallel(
  uconf = ucConf(),
  tdir = tempdir(),
  wtaq = system.file("extdata", "wtaq.2.1.exe", package = "kwb.wtaq"),
  FUN = NULL
)
}
\arguments{
\item{uconf}{UCode configuration as retrieved by \code{\link{ucConf}}}

\item{tdir}{tdir}

\item{wtaq}{wtaq}

\item{FUN}{function that is invoked each time before WTAQ is run}
}
\value{
???
}
\description{
Run main calibration loop
}
\seealso{
\code{\link{ucPrepCalib}}
}
