% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucSetPlaceholder}
\alias{ucSetPlaceholder}
\title{Set placeholder in WTAQ parameter setting}
\usage{
ucSetPlaceholder(parname, wtaqConfiguration, pchar = "@")
}
\arguments{
\item{parname}{Parameter name. Must be one of the aquifer parameters "bb",
"hkr", "hkz", "ss", "sy" (for description see
\code{\link{wtConfigureAquifer}}) or one of the drainage parameters "acc",
"akk", "amm", "axmm" (for description see
\code{\link{wtConfigureDrainage}}; parameter "alpha" currently not
supported) or one of the pumpwell parameters "qq", "rw", "rc", "zpd",
"zpl", "sw" (for description see \code{\link{wtConfigurePumpwell}}) or one
of the observation well parameters "r_i", "z1_i", "z2_i", "zp_i",
"rp_i", "xll_i" (for description see
\code{\link{wtConfigureObservationWell}}) where "i" is the number of the
observation well according to the list index in the "obswell" section of
the parameter setting.}

\item{wtaqConfiguration}{WTAQ parameter setting as created by means of
\code{\link{wtConfigure}}}

\item{pchar}{character to be used as the very first and very last character
of the placeholder. Default: \dQuote{@}}
}
\value{
Parameter setting in which a parameter value is replaced by an
appropriate placeholder
}
\description{
Set placeholder for parameter value in WTAQ parameter setting
}
\seealso{
\code{\link{ucConfParameter}, \link{ucConf}}
}
