% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucTable}
\alias{ucTable}
\title{Compose Table Data}
\usage{
ucTable(dat)
}
\arguments{
\item{dat}{data frame}
}
\description{
Creates lines to be put in TABLE-data block of UCODE input file representing
data in given data frame
}
