% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucTestBatchWtaq}
\alias{ucTestBatchWtaq}
\title{Test run of batch file}
\usage{
ucTestBatchWtaq(bfile, wtaqConfiguration)
}
\arguments{
\item{bfile}{path to batch file}

\item{wtaqConfiguration}{WTAQ configuration}
}
\description{
Test run of batch file
}
