% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucWriteBatchUcode}
\alias{ucWriteBatchUcode}
\title{Write batch file invoking UCODE}
\usage{
ucWriteBatchUcode(
  bfile,
  ucode = system.file("extdata", "ucode_2005.exe", package = "kwb.wtaq"),
  waitForInput = TRUE,
  inpfile = "ucode.in"
)
}
\arguments{
\item{bfile}{full path to batch file}

\item{ucode}{full path to UCODE executable file}

\item{waitForInput}{if \code{TRUE}, "pause" is added to the batch file}

\item{inpfile}{name of \code{ucode} input file}
}
\description{
Write batch file invoking UCODE
}
