% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucWriteBatchWtaq}
\alias{ucWriteBatchWtaq}
\title{Write batch file invoked by UCODE, running first WTAQ and second the R
script doing the goodness of fit evaluation}
\usage{
ucWriteBatchWtaq(bfile, wtaq, verbose, waitForInput = FALSE)
}
\arguments{
\item{bfile}{path to batch file}

\item{wtaq}{path to WTAQ executable}

\item{verbose}{if \code{TRUE}, the output is not discarded}

\item{waitForInput}{if \code{TRUE}, "pause" is added to the batch file}
}
\description{
Write batch file invoked by UCODE, running first WTAQ and second the R
script doing the goodness of fit evaluation
}
