% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucWriteBatchWtaqParallel}
\alias{ucWriteBatchWtaqParallel}
\title{Write batch file invoked by UCODE, which is just waiting until WTAQ has
finished and goodness of fit is available}
\usage{
ucWriteBatchWtaqParallel(bfile, recheckAfterMs = 10)
}
\arguments{
\item{bfile}{path to batchfile}

\item{recheckAfterMs}{time to wait for file, in milliseconds}
}
\description{
Write batch file invoked by UCODE, which is just waiting until WTAQ has
finished and goodness of fit is available
}
