% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucWriteEvalGofR}
\alias{ucWriteEvalGofR}
\title{Write R script evaluating goodness of fit from WTAQ output file}
\usage{
ucWriteEvalGofR(rfile, uconf)
}
\arguments{
\item{rfile}{path to R script file to be generated}

\item{uconf}{UCODE configuration}
}
\description{
Write R script evaluating goodness of fit from WTAQ output file
}
