% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucWriteInstruction}
\alias{ucWriteInstruction}
\title{Write UCODE instruction file, describing how to read goodness of fit values
from wtaq.gof}
\usage{
ucWriteInstruction(ifile, wtaqConfiguration, uconf)
}
\arguments{
\item{ifile}{path to instruction file}

\item{wtaqConfiguration}{WTAW configuration}

\item{uconf}{UCODE configuration}
}
\description{
Write UCODE instruction file, describing how to read goodness of fit values
from wtaq.gof
}
