% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucWriteMainInput}
\alias{ucWriteMainInput}
\title{Write UCODE main input file}
\usage{
ucWriteMainInput(ufile, uconf, command)
}
\arguments{
\item{ufile}{path to UCODE input file}

\item{uconf}{UCODE configuration}

\item{command}{name of batch file that is invoked by UCODE with each iteration}
}
\description{
Write UCODE main input file
}
