% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{ucWriteTemplate}
\alias{ucWriteTemplate}
\title{Write WTAQ input template file}
\usage{
ucWriteTemplate(tfile, wtaqConfiguration, uconf, pchar = "@")
}
\arguments{
\item{tfile}{path to template file}

\item{wtaqConfiguration}{WTAQ configuration}

\item{uconf}{UCODE configuration}

\item{pchar}{character indicating a placeholder}
}
\description{
Write WTAQ input template file
}
