% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{uwGofEval}
\alias{uwGofEval}
\title{Evaluate goodness of fit for WTAQ result}
\usage{
uwGofEval(res, logtimes = FALSE, gofs = names(uwGofTargetValue()), ...)
}
\arguments{
\item{res}{either character string representing full path to WTAQ result plot file or
data.frame containing the content of the WTAQ result plot file.}

\item{logtimes}{if TRUE, time steps are supposed to be logarithmic, i.e. it is assumed
that the result plot file is in \dQuote{matrix form}.}

\item{gofs}{vector of names of fitness functions, evaluated by \code{\link[hydroGOF]{gof}}.
Default: \code{c("ME", "MAE", "MSE", "RMSE", "NRMSE", "PBIAS", "RSR", 
  "rSD", "NSE", "mNSE", "rNSE", "d", "md", "rd", "cp", "r", "R2", "bR2", 
  "KGE", "VE")}}

\item{\dots}{arguments passed to gof() function, e.g. digits: decimal places used for
rounding the goodness-of-fit indexes.}
}
\value{
Evaluation matrix
}
\description{
Evaluate goodness of fit for WTAQ result
}
