% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtUcode.R
\name{uwGofTargetValue}
\alias{uwGofTargetValue}
\title{Target values for different GOF functions}
\usage{
uwGofTargetValue(gofs = uwGofNames())
}
\arguments{
\item{gofs}{names of GOF functions of which target values are to be returned}
}
\description{
Target values for different GOF functions, i.e. values that GOF functions
return if observed and calculated timeseries are identical
}
